Moving UpgradeSave and AddToLootTables to PlayerPostLoadPatch due to those being only referenced in that class
