+++ CraftMagicItems/Main.cs
@@ -127,7 +127,7 @@ namespace CraftMagicItems
 #endif
 
         public static bool modEnabled = true;
-        private static CraftMagicItemsBlueprintPatcher blueprintPatcher;
+        public static CraftMagicItemsBlueprintPatcher blueprintPatcher;
 
         public static readonly Dictionary<ItemEntity, CraftingProjectData> ItemUpgradeProjects = new Dictionary<ItemEntity, CraftingProjectData>();
         public static readonly List<CraftingProjectData> ItemCreationProjects = new List<CraftingProjectData>();
@@ -3195,59 +3195,6 @@ namespace CraftMagicItems
             if (daysAvailableToCraft > 0) {
                 // They didn't use up all available days - reset the time they can start crafting to now.
                 timer.LastUpdated = Game.Instance.Player.GameTime;
-            }
-        }
-
-        private static void AddToLootTables(BlueprintItem blueprint, string[] tableNames, bool firstTime) {
-            var tableCount = tableNames.Length;
-            foreach (var loot in ResourcesLibrary.GetBlueprints<BlueprintLoot>()) {
-                if (tableNames.Contains(loot.name)) {
-                    tableCount--;
-                    if (!loot.Items.Any(entry => entry.Item == blueprint)) {
-                        var lootItems = loot.Items.ToList();
-                        lootItems.Add(new LootEntry {Count = 1, Item = blueprint});
-                        loot.Items = lootItems.ToArray();
-                    }
-                }
-            }
-            foreach (var unitLoot in ResourcesLibrary.GetBlueprints<BlueprintUnitLoot>()) {
-                if (tableNames.Contains(unitLoot.name)) {
-                    tableCount--;
-                    if (unitLoot is BlueprintSharedVendorTable vendor) {
-                        if (firstTime) {
-                            var vendorTable = Game.Instance.Player.SharedVendorTables.GetTable(vendor);
-                            vendorTable.Add(blueprint.CreateEntity());
-                        }
-                    } else if (!unitLoot.ComponentsArray.Any(component => component is LootItemsPackFixed pack && pack.Item.Item == blueprint)) {
-                        var lootItem = new LootItem();
-                        Accessors.SetLootItemItem(lootItem) = blueprint;
-#if PATCH21_BETA
-                        var lootComponent = SerializedScriptableObject.CreateInstance<LootItemsPackFixed>();
-#else
-                        var lootComponent = ScriptableObject.CreateInstance<LootItemsPackFixed>();
-#endif
-                        Accessors.SetLootItemsPackFixedItem(lootComponent) = lootItem;
-                        blueprintPatcher.EnsureComponentNameUnique(lootComponent, unitLoot.ComponentsArray);
-                        var components = unitLoot.ComponentsArray.ToList();
-                        components.Add(lootComponent);
-                        unitLoot.ComponentsArray = components.ToArray();
-                    }
-                }
-            }
-            if (tableCount > 0) {
-                HarmonyLib.FileLog.Log($"!!! Failed to match all loot table names for {blueprint.Name}.  {tableCount} table names not found.");
-            }
-        }
-
-        public static void UpgradeSave(Version version) {
-            foreach (var lootItem in LoadedData.CustomLootItems) {
-                var firstTime = (version == null || version.CompareTo(lootItem.AddInVersion) < 0);
-                var item = ResourcesLibrary.TryGetBlueprint<BlueprintItem>(lootItem.AssetGuid);
-                if (item == null) {
-                    HarmonyLib.FileLog.Log($"!!! Loot item not created: {lootItem.AssetGuid}");
-                } else {
-                    AddToLootTables(item, lootItem.LootTables, firstTime);
-                }
             }
         }
